pyproject.toml
[tool.poetry]
name = "export-hunter"
version = "0.1.0"
description = ""
authors = ["Scareface <tektonikboy98@gmail.com>"]
readme = "README.md"
packages = [{include = "export_hunter"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.0.2"
comtradeapicall = "^1.0.20"
eurostat = "^1.0.4"
django = "^4.2.2"
psycopg = "^3.1.9"
django-cors-headers = "^4.0.0"
djangorestframework = "^3.14.0"
drf-spectacular = {extras = ["sidecar"], version = "^0.26.2"}
djangorestframework-simplejwt = "^5.2.2"
python-dotenv = "^1.0.0"
flake8 = "^6.0.0"
django-rest-passwordreset = "^1.3.0"
coverage = "^7.2.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
